Fixed the file structure for each js files to their separate folders
