+++ app/javascript/components/(NotUsedATM)HomePage.js
