+++ app/javascript/components/App.js
@@ -1,9 +1,8 @@
 import React from "react"
 import PropTypes from "prop-types"
-import HomePage from "./HomePage"
-import CareerPage from './CareerPage'
-import CreateJob from './CreateJob'
-import ShowCurrentJob from './ShowCurrentJob'
+import CareerMainPage from './CareerMainPage'
+import CreateJob from './pages/jobs/CreateJob'
+import ShowCurrentJob from './pages/jobs/ShowCurrentJob'
 
 import Header from "./pages/Header"
 import { BrowserRouter as Router, Link, Switch, Route } from "react-router-dom"
@@ -20,14 +19,6 @@ function App(props) {
         <Router>
           <React.Fragment>
 
-            <Switch>
-                {!logged_in &&
-                    <Route path="/">
-                        <HomePage />
-                    </Route>
-                }
-            </Switch>
-
             <Header
                 logged_in={logged_in}
                 sign_in_route={sign_in_route}
@@ -38,14 +29,14 @@ function App(props) {
                 {logged_in &&
                     <div>
                         <Switch>
-                            <Route exact path="/careerpage">
-                                <CareerPage
+                            <Route exact path="/careermainpage">
+                                <CareerMainPage
                                     current_user_id={current_user_id}
                                  />
                             </Route>
 
                             <Route exact path='/jobs/:id' >
-                                <ShowCurrentJob /> 
+                                <ShowCurrentJob />
                             </Route>
